changing it to u128 makes wasmbuild/wasm_bindgen complain
[btwiuse@railway wasmbuild]$ make
deno task wasmbuild
Task wasmbuild deno run -A https://deno.land/x/wasmbuild@0.15.3/main.ts
Ensuring wasm32-unknown-unknown target installed...
Building wasmbuild WebAssembly...
  build --lib -p wasmbuild --target wasm32-unknown-unknown --release
   Compiling wasmbuild v0.0.4 (/home/btwiuse/wasmbuild)
error[E0277]: the trait bound `u128: FromWasmAbi` is not satisfied
 --> lib.rs:4:1
  |
4 | #[wasm_bindgen]
  | ^^^^^^^^^^^^^^^ the trait `FromWasmAbi` is not implemented for `u128`
  |
  = help: the following other types implement trait `FromWasmAbi`:
            isize
            i8
            i16
            i32
            i64
            usize
            u8
            u16
          and 4 others
  = note: this error originates in the attribute macro `wasm_bindgen` (in Nightly builds, run with -Z macro-backtrace for more info)

error[E0277]: the trait bound `u128: IntoWasmAbi` is not satisfied
 --> lib.rs:4:1
  |
4 | #[wasm_bindgen]
  | ^^^^^^^^^^^^^^^ the trait `IntoWasmAbi` is not implemented for `u128`
  |
  = help: the following other types implement trait `IntoWasmAbi`:
            isize
            i8
            i16
            i32
            i64
            usize
            u8
            u16
          and 4 others
  = note: required for `u128` to implement `ReturnWasmAbi`
  = note: this error originates in the attribute macro `wasm_bindgen` (in Nightly builds, run with -Z macro-backtrace for more info)

For more information about this error, try `rustc --explain E0277`.
error: could not compile `wasmbuild` (lib) due to 2 previous errors
cargo build failed
make: *** [Makefile:2: all] Error 1

+++ lib.rs
@@ -1,8 +1,8 @@
 use wasm_bindgen::prelude::*;
 
-/// add two u64 numbers
+/// add two u128 numbers
 #[wasm_bindgen]
-pub fn add(a: u64, b: u64) -> u64 {
+pub fn add(a: u128, b: u128) -> u128 {
   return a + b;
 }
